Renamed ssh-python packages, moved base clients to own package. Moved client tests into packages
