+++ .travis.yml
@@ -28,13 +28,9 @@ install:
 script:
   # For testing SSH agent related functionality
   - eval `ssh-agent -s`
-  - pytest --reruns 5 --cov-append --cov=pssh tests/test_native_tunnel.py
-  - pytest --reruns 5 --cov-append --cov=pssh tests/test_utils.py
-  - pytest --reruns 5 --cov-append --cov=pssh tests/test_native_*_client.py
-  - pytest --reruns 5 --cov-append --cov=pssh tests/test_paramiko_*.py
-  - pytest --reruns 5 --cov-append --cov=pssh tests/ssh_lib
-  - pytest --cov-append --cov=pssh tests/test_imports.py
-  - pytest --cov-append --cov=pssh tests/test_output.py
+  - pytest --cov-append --cov=pssh tests/test_imports.py tests/test_output.py tests/test_utils.py
+  - pytest --reruns 5 --cov-append --cov=pssh tests/miko
+  - pytest --reruns 5 --cov-append --cov=pssh tests/native tests/ssh
   - flake8 pssh
   - cd doc; make html; cd ..
   # Test building from source distribution
