+++ pssh/clients/base/parallel.py
@@ -27,9 +27,9 @@ from warnings import warn
 from gevent import joinall
 from gevent.hub import Hub
 
-from ..exceptions import HostArgumentException
-from ..constants import DEFAULT_RETRIES, RETRY_DELAY
-from ..output import HostOutput
+from ...constants import DEFAULT_RETRIES, RETRY_DELAY
+from ...exceptions import HostArgumentException
+from ...output import HostOutput
 
 
 Hub.NOT_ERROR = (Exception,)
