+++ pssh/clients/ssh/__init__.py
@@ -1,0 +1,20 @@
+# This file is part of parallel-ssh.
+#
+# Copyright (C) 2014-2020 Panos Kittenis.
+#
+# This library is free software; you can redistribute it and/or
+# modify it under the terms of the GNU Lesser General Public
+# License as published by the Free Software Foundation, version 2.1.
+#
+# This library is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public
+# License along with this library; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+
+# flake8: noqa: F401
+from .parallel import ParallelSSHClient
+from .single import SSHClient
