+++ tests/embedded_server/sshd_config.tmpl
@@ -8,4 +8,4 @@ AcceptEnv LANG LC_*
 Subsystem sftp internal-sftp
 AuthorizedKeysFile {{parent_dir}}/authorized_keys
 MaxSessions 100
-PidFile {{random_server}}.pid
+PidFile {{parent_dir}}/{{random_server}}.pid
